Refactor input/prefs and model/eval
