+++ lib/model/eval.js
@@ -21,33 +21,46 @@ const getRequestArray = (evalOptions) => {
   return [idNo, lab, commitHash, lang];
 };
 
+const socketEventHandler = (event, socket, data, callback) => {
+  let response;
+  switch (event) {
+    case 'invalid':
+      logger.moduleLog('error', 'Eval Model', 'Invalid options provided.');
+      response = {
+        name: 'invalid',
+      };
+      break;
+    case 'submission_pending':
+      logger.moduleLog('debug', 'Eval Model', 'Submission Pending...');
+      response = {
+        name: 'submission_pending',
+      };
+      break;
+    case 'scores':
+      logger.moduleLog('debug', 'Eval Model', 'Sucessfully Sumbitted. Evaluated results');
+      response = {
+        name: 'scores',
+        details: {
+          ...data,
+        },
+      };
+      break;
+  }
+
+  logger.moduleLog('error', 'Eval Model', data);
+  socket.close();
+  callback(response);
+};
+
 const getResponse = (socket, callback) => {
   socket.on('invalid', (data) => {
-    logger.moduleLog('error', 'Eval Model', 'Invalid options provided.');
-    logger.moduleLog('error', 'Eval Model', data);
-    socket.close();
-    callback({
-      name: 'invalid',
-    });
+    socketEventHandler('invalid', socket, data, callback);
   });
   socket.on('submission_pending', (data) => {
-    logger.moduleLog('debug', 'Eval Model', 'Submission Pending...');
-    logger.moduleLog('debug', 'Eval Model', data);
-    socket.close();
-    callback({
-      name: 'submission_pending',
-    });
+    socketEventHandler('submission_pending', socket, data, callback);
   });
   socket.on('scores', (data) => {
-    logger.moduleLog('debug', 'Eval Model', 'Sucessfully Sumbitted. Evaluated results');
-    logger.moduleLog('debug', 'Eval Model', data);
-    socket.close();
-    callback({
-      name: 'scores',
-      details: {
-        ...data,
-      },
-    });
+    socketEventHandler('scores', socket, data, callback);
   });
 };
